added some practical programs

+++ Lab Work/overloading.java
@@ -3,14 +3,15 @@ public class overloading {
         return n1 + n2;
     }
 
-    double sum(double n1, double n2) {
+    float
+     sum(float n1, float n2) {
         return n1 + n2;
     }
 
     public static void main(String[] args) {
         overloading o = new overloading();
         System.out.println(o.sum(4, 5));
-        System.out.println(o.sum(4.5, 5.5));
+        System.out.println(o.sum(4.5f, 5.5f));
 
     }
 }
